Share server runtime source closure
